dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile dev-requirements.in
#
anyio==3.3.1
    # via jupyter-server
argon2-cffi==21.1.0
    # via
    #   jupyter-server
    #   notebook
attrs==21.2.0
    # via jsonschema
autopep8==1.5.7
    # via -r dev-requirements.in
babel==2.9.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
backports.entry-points-selectable==1.1.0
    # via virtualenv
bleach==4.1.0
    # via nbconvert
certifi==2021.5.30
    # via
    #   -c requirements.txt
    #   requests
cffi==1.14.6
    # via argon2-cffi
charset-normalizer==2.0.6
    # via
    #   -c requirements.txt
    #   requests
click==8.0.1
    # via pip-tools
colorama==0.4.4
    # via
    #   -c requirements.txt
    #   click
    #   ipython
debugpy==1.4.3
    # via ipykernel
decorator==5.1.0
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distlib==0.3.3
    # via virtualenv
entrypoints==0.3
    # via
    #   jupyter-client
    #   jupyterlab-server
    #   nbconvert
filelock==3.0.12
    # via virtualenv
idna==3.2
    # via
    #   -c requirements.txt
    #   anyio
    #   requests
ipykernel==6.4.1
    # via
    #   ipywidgets
    #   notebook
ipython==7.27.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-server
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via -r dev-requirements.in
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via scikit-learn
json5==0.9.6
    # via jupyterlab-server
jsonschema==3.2.0
    # via
    #   jupyterlab-server
    #   nbformat
jupyter-client==7.0.3
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
    #   notebook
jupyter-core==4.8.1
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-server==1.11.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
jupyterlab==3.1.13
    # via -r dev-requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-server==2.8.1
    # via jupyterlab
jupyterlab-widgets==1.0.2
    # via ipywidgets
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclassic==0.3.2
    # via jupyterlab
nbclient==0.5.4
    # via nbconvert
nbconvert==6.1.0
    # via
    #   jupyter-server
    #   notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
notebook==6.4.4
    # via
    #   nbclassic
    #   widgetsnbextension
numpy==1.21.2
    # via
    #   -c requirements.txt
    #   scikit-learn
    #   scipy
packaging==21.0
    # via
    #   bleach
    #   jupyterlab
    #   jupyterlab-server
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.2
    # via jedi
pep517==0.11.0
    # via pip-tools
pickleshare==0.7.5
    # via ipython
pip-tools==6.3.0
    # via -r dev-requirements.in
platformdirs==2.3.0
    # via virtualenv
prometheus-client==0.11.0
    # via
    #   jupyter-server
    #   notebook
prompt-toolkit==3.0.20
    # via ipython
pycodestyle==2.7.0
    # via autopep8
pycparser==2.20
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pymysql==1.0.2
    # via -r dev-requirements.in
pyparsing==2.4.7
    # via
    #   -c requirements.txt
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   -c requirements.txt
    #   jupyter-client
pytz==2021.1
    # via
    #   -c requirements.txt
    #   babel
pywin32==301
    # via jupyter-core
pywinpty==1.1.4
    # via terminado
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   notebook
requests==2.26.0
    # via
    #   -c requirements.txt
    #   jupyterlab-server
    #   requests-unixsocket
requests-unixsocket==0.2.0
    # via jupyter-server
scikit-learn==1.0
    # via sklearn
scipy==1.7.1
    # via
    #   -c requirements.txt
    #   scikit-learn
send2trash==1.8.0
    # via
    #   jupyter-server
    #   notebook
six==1.16.0
    # via
    #   -c requirements.txt
    #   bleach
    #   jsonschema
    #   python-dateutil
    #   virtualenv
sklearn==0.0
    # via -r dev-requirements.in
sniffio==1.2.0
    # via anyio
terminado==0.12.1
    # via
    #   jupyter-server
    #   notebook
testpath==0.5.0
    # via nbconvert
threadpoolctl==3.0.0
    # via scikit-learn
toml==0.10.2
    # via autopep8
tomli==1.2.1
    # via pep517
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.1.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
urllib3==1.26.7
    # via
    #   -c requirements.txt
    #   requests
    #   requests-unixsocket
virtualenv==20.8.0
    # via -r dev-requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websocket-client==1.2.1
    # via jupyter-server
wheel==0.37.0
    # via pip-tools
widgetsnbextension==3.5.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
